+++ x/leverage/types/flag.go
@@ -1,0 +1,3 @@
+package types
+
+const FlagEnableLiquidatorQuery = "enable-liquidator-query"